chore: move FlowKeyboardPreview to its own component, fix a typo and remove test.  Will create issue for the test.
